requirements.txt
asyncio==3.4.3
discord.py==1.7.2
Flask==1.1.2
slackclient==2.9.3
slackeventsapi==2.2.1